REFACTOR: (SPOT-166) Route와 Event의 dto, implement 내에서 분리한다
