+++ src/main/java/com/meetup/server/event/dto/response/route/DrivingInfoResponse.java
@@ -1,4 +1,4 @@
-package com.meetup.server.event.dto.response;
+package com.meetup.server.event.dto.response.route;
 
 import com.meetup.server.global.clients.kakao.mobility.KakaoMobilityResponse;
 import lombok.Builder;
