+++ src/main/java/com/meetup/server/event/implement/EventProcessor.java
@@ -3,7 +3,7 @@ package com.meetup.server.event.implement;
 import com.meetup.server.event.domain.Event;
 import com.meetup.server.event.dto.request.EventRequest;
 import com.meetup.server.event.dto.request.UpdateEventRequest;
-import com.meetup.server.event.dto.response.RouteResponse;
+import com.meetup.server.event.dto.response.route.RouteResponse;
 import com.meetup.server.event.infrastructure.jpa.EventRepository;
 import com.meetup.server.review.implement.ReviewWriter;
 import com.meetup.server.startpoint.implement.StartPointProcessor;
