+++ src/main/java/com/meetup/server/event/implement/route/MeetingPointCalculator.java
@@ -1,9 +1,11 @@
-package com.meetup.server.event.implement;
+package com.meetup.server.event.implement.route;
 
 import com.meetup.server.event.domain.Event;
-import com.meetup.server.event.dto.response.MeetingPointResult;
+import com.meetup.server.event.dto.response.route.MeetingPointResult;
 import com.meetup.server.event.exception.EventErrorType;
 import com.meetup.server.event.exception.EventException;
+import com.meetup.server.event.implement.EventReader;
+import com.meetup.server.event.implement.EventValidator;
 import com.meetup.server.global.util.CoordinateUtil;
 import com.meetup.server.startpoint.domain.StartPoint;
 import com.meetup.server.startpoint.implement.StartPointReader;
