+++ src/main/java/com/meetup/server/event/implement/route/RouteApiCaller.java
@@ -1,4 +1,4 @@
-package com.meetup.server.event.implement;
+package com.meetup.server.event.implement.route;
 
 import com.meetup.server.global.clients.kakao.mobility.KakaoMobilityClient;
 import com.meetup.server.global.clients.kakao.mobility.KakaoMobilityRequest;
